Separate views.py to python package
