+++ backend/accounts/views/api.py
@@ -1,0 +1,2 @@
+class MovieAPIView:
+    pass
