+++ backend/movie/views/__init__.py
@@ -1,0 +1,5 @@
+from .api import MovieAPIView
+
+from .list import (
+    MovieListView, ActorListView, DirectorListView
+)
