+++ backend/movie/views/api.py
@@ -1,0 +1,13 @@
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework import status
+
+from movie.models import Movie
+from movie.serializers import MovieSerializer
+
+
+class MovieAPIView(APIView):
+    def get(self, request):
+        query_set = Movie.objects.all()
+        serializer = MovieSerializer(query_set, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
