+++ backend/movie/views/list.py
@@ -1,11 +1,6 @@
-from django.shortcuts import render
 from django.views.generic import ListView
-from rest_framework.response import Response
-from rest_framework.views import APIView
-from rest_framework import status
 
-from .models import Movie, Actor, Director
-from .serializers import MovieSerializer
+from movie.models import Movie, Actor, Director
 # Create your views here.
 
 
@@ -28,10 +23,3 @@ class DirectorListView(ListView):
     context_object_name = 'persons'
     template_name = 'movie/person_list.html'
     paginate_by = 12
-
-
-class MovieAPIView(APIView):
-    def get(self, request):
-        query_set = Movie.objects.all()
-        serializer = MovieSerializer(query_set, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
